Adding leafs iter
Added leafs iter from fap4. Leaf here is pair.
So leaf iter of tree node is actually iter of its binded node.
With such scheme there is the problem in ifr - we need to replace the
functionality that currently provided by firstLeafB() override that
allows ifr to resolve invalid nodes when the node is included in
iteration path. The idea is to use leafsBegin() but actually
during traversal the binded leafsBegin() is called instead of tree
node leafsBegin(). Needs to improve the scheme.
In the current commit ifr doesn't perform such resoution by demand
so is not working properly.